requirements.txt
opencv-python
Pillow
torch
torchvision
tqdm
numpy
scipy
pyclipper
shapely
natsort
python-magic
docopt-ng
ConfigUpdater
manga_ocr
logzero
prettytable
colorama
requests
xdg